+++ lib/lottie/exporters/svg.py
@@ -2,6 +2,7 @@ from xml.dom import minidom
 from xml.etree import ElementTree
 
 from .base import exporter
+from ..parsers.baseporter import ExtraOption
 from ..parsers.svg.builder import to_svg
 from ..utils.file import open_file
 
@@ -16,7 +17,9 @@ def _print_pretty_xml(dom, file):
         fp.write(xmlstr)
 
 
-@exporter("SVG", ["svg"], [], {"pretty", "frame"})
-def export_svg(animation, file, frame=0, pretty=True):
+@exporter("SVG", ["svg"], [
+    ExtraOption("animated", action="store_true", help="Export animated svg"),
+], {"pretty", "frame"})
+def export_svg(animation, file, frame=0, pretty=True, animated=False):
     _print_xml = _print_pretty_xml if pretty else _print_ugly_xml
-    _print_xml(to_svg(animation, frame), file)
+    _print_xml(to_svg(animation, frame, animated), file)