+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Unreleased
  * Fixed font fallback
+ * Optional glaxnimate integration
+ * Animated SVG output
 
 ## 0.6.7 - 2021-03-23
  * Improved text layer support